backend/requirements.txt
teler==0.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pydub
webrtcvad
numpy
psycopg2-binary
pgvector
voyageai
python-multipart
PyPDF2
python-docx
openpyxl
langchain
langchain-community
tiktoken